refactor: ModelIdを別ファイルに抽出
